hotfix: в Navbar используется RouterPaths
