+++ packages/client/src/shared/router/index.ts
@@ -1,0 +1,2 @@
+export { RouterPaths } from './RouterPaths'
+export { router } from './router'
