+++ packages/client/src/shared/router/router.tsx
@@ -1,4 +1,5 @@
 import type { RouteProps } from 'react-router-dom'
+import { AppRoutes, RouterPaths, type AppRoutesValues } from './RouterPaths'
 
 import {
   NotFoundPage,
@@ -13,37 +14,6 @@ import {
   ErrorPage,
   EditPasswordPage,
 } from '@/pages'
-
-export const AppRoutes = {
-  MAIN: 'main',
-  LOGIN: 'login',
-  REGISTRATION: 'registration',
-  PASSWORD: 'edit-password',
-  PROFILE: 'profile',
-  LEADERBOARD: 'leaderboard',
-  GAME: 'game',
-  FORUM: 'forum',
-  TOPIC: 'topic',
-  NOTFOUND: 'not-found',
-  ERROR: 'error',
-} as const
-
-type AppRoutesKeys = keyof typeof AppRoutes
-type AppRoutesValues = typeof AppRoutes[AppRoutesKeys]
-
-export const RouterPaths: Record<AppRoutesValues, string> = {
-  [AppRoutes.MAIN]: '/',
-  [AppRoutes.LOGIN]: '/sign-in',
-  [AppRoutes.REGISTRATION]: '/sign-up',
-  [AppRoutes.PROFILE]: '/profile',
-  [AppRoutes.PASSWORD]: '/edit-password',
-  [AppRoutes.GAME]: '/game',
-  [AppRoutes.LEADERBOARD]: '/leaderboard',
-  [AppRoutes.FORUM]: '/forum',
-  [AppRoutes.NOTFOUND]: '*',
-  [AppRoutes.TOPIC]: '/topic/:topicId',
-  [AppRoutes.ERROR]: '/error',
-}
 
 export const router: Record<AppRoutesValues, RouteProps> = {
   [AppRoutes.MAIN]: {
